refactor: move common tpes to file
