--- src/types/interfaces.ts
@@ -1,9 +1,0 @@
-export interface Movie {
-  id: string;
-  name: string;
-  genres: string[];
-  year: number;
-  poster: string;
-  runtime: number;
-  description: string
-}
